restructure and adding tests
